Lab 4: JavaScript Clue Game -- Mixing objects and arrays
Do you remember the classic detective board game Clue? It was a very popular game in the 90's.

The plot starts with the death of Mr. Boddy, the very wealthy owner of Tudor Manor. He was killed under highly suspicious circumstances, and it's the player's mission to uncover which suspect murdered him, what weapon was used, and in which room of the mansion the crime took place.

When playing Clue, these three details are represented by a suspect card, a weapon card and a room card that are kept hidden in an envelope for the duration of the game. By progressively revealing the remaining cards, the players can infer which are the three cards that are hidden. When a player is confident that they know the solution to the mystery, they will try to reveal it. If a player guesses correctly, they win the game.


Step 1: Create the cards

Clue has three different card types: suspects, rooms, and weapons.

Let's create an array for every card type. These arrays should be named suspectsArray, weaponsArray, roomsArray. Each array will contain strings, where each string represents one card.

All of the information you need about suspects, rooms, and weapons can be made up. You should have the following amounts of each card type:

Suspects: 6

Weapons: 8

Rooms: 10


Step 2: Create the mystery

At the beginning of the game, players shuffle each of the card stacks to create a combination of suspect, weapon and room. This will be the mystery to solve.


Random selector

Declare a function named selectRandom to randomly select one element from a card stack. The function should expect an array as an argument and should return a random element from the array.


Create the mystery

Declare a function named pickMystery that takes no arguments and returns an object with three properties: suspect, weapon and room, each holding as a value a card of that specific type. You can get a random card of each type by calling selectRandom on each card stack.


Step 3: Reveal the mystery

At last, we are ready to reveal the mystery.

Declare a function named revealMystery that receives the object returned by pickMystery as the single argument, and returns a reveal message in the following format:

<NAME> killed Mr. Marist using the <WEAPON> in the <ROOM>!